requirements.txt
mysql-connector
gunicorn==20.1.0